move _mexpand to function; give recursive option
The expressions from test_roots_slow in test_polyroots were
moved to test_arit where their ability to simplify to zero
is tested and used to demonstrate the need for a recursive
flag on _mexpand.
