+++ sympy/solvers/bivariate.py
@@ -10,7 +10,8 @@ from sympy.core.symbol import (Dummy, Wild)
 from sympy.functions.elementary.exponential import (LambertW, exp, log)
 from sympy.functions.elementary.miscellaneous import root
 from sympy.polys.polytools import (Poly, primitive, factor)
-from sympy.simplify.simplify import (_mexpand, collect, separatevars)
+from sympy.core.function import _mexpand
+from sympy.simplify.simplify import (collect, separatevars)
 from sympy.solvers.solvers import solve, _invert
